+++ core/chains/evm/config/chain_specific_config.go
@@ -59,7 +59,6 @@ type (
 		maxQueuedTransactions                          uint64
 		minGasPriceWei                                 big.Int
 		minIncomingConfirmations                       uint32
-		minRequiredOutgoingConfirmations               uint64
 		minimumContractPayment                         *assets.Link
 		nodeDeadAfterNoNewHeadersThreshold             time.Duration
 		nodePollFailureThreshold                       uint32
@@ -136,7 +135,6 @@ func setChainSpecificConfigDefaultSets() {
 		maxQueuedTransactions:                 250,
 		minGasPriceWei:                        *assets.GWei(1),
 		minIncomingConfirmations:              3,
-		minRequiredOutgoingConfirmations:      12,
 		minimumContractPayment:                DefaultMinimumContractPayment,
 		nodeDeadAfterNoNewHeadersThreshold:    3 * time.Minute,
 		nodePollFailureThreshold:              5,
@@ -206,7 +204,6 @@ func setChainSpecificConfigDefaultSets() {
 	bscMainnet.linkContractAddress = "0x404460c6a5ede2d891e8297795264fde62adbb75"
 	bscMainnet.minGasPriceWei = *assets.GWei(1)
 	bscMainnet.minIncomingConfirmations = 3
-	bscMainnet.minRequiredOutgoingConfirmations = 12
 	bscMainnet.ocrDatabaseTimeout = 2 * time.Second
 	bscMainnet.ocrContractTransmitterTransmitTimeout = 2 * time.Second
 	bscMainnet.ocrObservationGracePeriod = 500 * time.Millisecond
@@ -234,7 +231,6 @@ func setChainSpecificConfigDefaultSets() {
 	polygonMainnet.blockHistoryEstimatorBlockHistorySize = 24
 	polygonMainnet.linkContractAddress = "0xb0897686c545045afc77cf20ec7a532e3120e0f1"
 	polygonMainnet.minIncomingConfirmations = 5
-	polygonMainnet.minRequiredOutgoingConfirmations = 12
 	polygonMainnet.logPollInterval = 1 * time.Second
 	polygonMumbai := polygonMainnet
 	polygonMumbai.gasPriceDefault = *assets.GWei(1)
@@ -271,7 +267,6 @@ func setChainSpecificConfigDefaultSets() {
 	optimismMainnet.linkContractAddress = "0x350a791Bfc2C21F9Ed5d10980Dad2e2638ffa7f6"
 	optimismMainnet.minIncomingConfirmations = 1
 	optimismMainnet.minGasPriceWei = *big.NewInt(0) // Optimism uses the Optimism2 estimator; we don't want to place any limits on the minimum gas price
-	optimismMainnet.minRequiredOutgoingConfirmations = 1
 	optimismMainnet.ocrContractConfirmations = 1
 	optimismKovan := optimismMainnet
 	optimismKovan.blockEmissionIdleWarningThreshold = 30 * time.Minute
@@ -283,7 +278,6 @@ func setChainSpecificConfigDefaultSets() {
 	fantomMainnet.maxGasPriceWei = *assets.GWei(200000)
 	fantomMainnet.linkContractAddress = "0x6f43ff82cca38001b6699a8ac47a2d0e66939407"
 	fantomMainnet.minIncomingConfirmations = 3
-	fantomMainnet.minRequiredOutgoingConfirmations = 2
 	fantomMainnet.logPollInterval = 1 * time.Second
 	fantomTestnet := fantomMainnet
 	fantomTestnet.linkContractAddress = "0xfafedb041c0dd4fa2dc0d87a6b0979ee6fa7af5f"
@@ -312,7 +306,6 @@ func setChainSpecificConfigDefaultSets() {
 	avalancheMainnet.blockHistoryEstimatorBlockHistorySize = 24 // Average block time of 2s
 	avalancheMainnet.blockHistoryEstimatorBlockDelay = 2
 	avalancheMainnet.minIncomingConfirmations = 1
-	avalancheMainnet.minRequiredOutgoingConfirmations = 1
 	avalancheMainnet.ocrContractConfirmations = 1
 	avalancheMainnet.logPollInterval = 3 * time.Second
 
@@ -324,7 +317,6 @@ func setChainSpecificConfigDefaultSets() {
 	harmonyMainnet.linkContractAddress = "0x218532a12a389a4a92fC0C5Fb22901D1c19198aA"
 	harmonyMainnet.gasPriceDefault = *assets.GWei(5)
 	harmonyMainnet.minIncomingConfirmations = 1
-	harmonyMainnet.minRequiredOutgoingConfirmations = 2
 	harmonyMainnet.logPollInterval = 2 * time.Second
 	harmonyTestnet := harmonyMainnet
 	harmonyTestnet.linkContractAddress = "0x8b12Ac23BFe11cAb03a634C1F117D64a7f2cFD3e"